+++ src/Event.php
@@ -214,7 +214,7 @@ class Event
     }
 
     /**
-     * Build the comand string.
+     * Build the command string.
      *
      * @return string
      */