+++ src/Task/Collection.php
@@ -39,6 +39,7 @@ class Collection
         $iterator = $this->finder
             ->files()
             ->name("*{$suffix}")
+            ->sortByName()
             ->in($source)
         ;